Add payment api and front page done

+++ src/main/java/com/jy/sb7/res/ReservationVO.java
@@ -1,19 +1,22 @@
 package com.jy.sb7.res;
 
+
 import com.jy.sb7.member.MemberVO;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @Data 
-public class ReservationVO extends MemberVO {
-	private int revNum;
-	private int startTime;
-	private int endTime;
+public class ReservationVO extends MemberVO{
+	private long revNum;
+	private long startTime;
+	private long endTime;
 	private int roomTime;
 	private int roomUser;
 	private String roomDate;
 	private int roomType;
 	private int roomPrice;
 	private String payment;
+	
 	
 }
